% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_maps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{HTML annotation for epicenter map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Filtered data frame with earthquakes to visualize.}
}
\value{
Character string for each earthquake, containing location (as cleaned
by \code{eq_location_clean}), magnitude and total number of deaths.
}
\description{
Creates an HTML label that can be used as the annotation text in
the leaflet earthquakes epicenters map from \code{eq_map}.
}
\examples{
\dontrun{
 noaase_raw <- readr::read_delim("inst/extdata/earthquakes.tsv.gz", "\\t")
 noaase_raw \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
   dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
   eq_map(annot_col = "popup_text")
}

}
