% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans location name column from raw NOAA data frame}
\usage{
eq_location_clean(noaase_raw)
}
\arguments{
\item{noaase_raw}{Raw NOAA data frame read with a function like
\code{readr::read_delim}.}
}
\value{
NOAA data frame, with LOCATION_NAME column cleaned by stripping out
the everything before the last colon (including the colon and space, usually
means country name, but there could be another subregion name) and converting
names to title case (as opposed to all caps).
}
\description{
Takes raw NOAA data frame and returns a data frame with
LOCATION_NAME column clean, by stripping out the country name (including the
colon) and converting names to title case (as opposed to all caps).
}
\examples{
\dontrun{
  noaase_raw <- readr::read_delim("inst/extdata/earthquakes.tsv.gz", "\\t")
  noaase_loc <- eq_location_clean(noaase_raw)
}

}
