% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_maps.R
\name{eq_map}
\alias{eq_map}
\title{Epicenter map with annotation}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{Filtered data frame with earthquakes to visualize.}

\item{annot_col}{Choose which column is used for the annotation in the
pop-up.}
}
\value{
Leaflet map with earthquakes epicenters marked and the contents in
\code{annot_col} as annotation in pop-up window.
}
\description{
Maps earthquakes epicenters and annotates each point with pop-up
window.
}
\examples{
\dontrun{
  noaase_raw <- readr::read_delim("inst/extdata/earthquakes.tsv.gz", "\\t")
  noaase_raw \%>\%
    eq_clean_data() \%>\%
    dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
    eq_map(annot_col = "DATE")
}

}
