% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Earthquake timeline plots}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Timeline ggplot2 geom for NOAA significant earthquakes database.
}
\section{Aesthetics}{

\code{geom_timeline} understands the following aesthetics (required
aesthetics are in bold):
\itemize{
  \item \strong{x}
  \item y
  \item alpha
  \item colour
  \item fill
  \item shape
  \item size
  \item stroke
}
}

\examples{
\dontrun{
noaase_raw <- readr::read_delim("inst/extdata/earthquakes.tsv.gz", "\\t")
noaase_raw \%>\%
  eq_clean_data() \%>\%
  filter(YEAR >= 2000, COUNTRY == "USA") \%>\%
  ggplot(ggplot2::aes(x = DATE,
                      size = as.numeric(EQ_PRIMARY),
                      color = as.numeric(TOTAL_DEATHS))) +
  geom_timeline() +
  theme_classic() +
  labs(size = "Richter scale value", color = "# deaths") +
  theme(axis.line.y = ggplot2::element_blank(),
        axis.ticks.y = ggplot2::element_blank(),
        axis.title.y = ggplot2::element_blank(),
        legend.position = "bottom") +
  ggplot2::guides(size = ggplot2::guide_legend(order = 1))
}

}
